refactor(StationEnvironmentElement): move classes
